Make some changes to get SQLite to pass the test suite

+++ tests/support.py
@@ -1,7 +1,7 @@
 import os
 
 from databases import Database
-from sqlalchemy import BigInteger, Column, ForeignKey, MetaData, String, Table
+from sqlalchemy import Integer, Column, ForeignKey, MetaData, String, Table
 from sqlalchemy.dialects import postgresql
 
 from datamapper import Associations, BelongsTo, HasMany, HasOne, Model
@@ -17,7 +17,7 @@ class User(Model):
     __table__ = Table(
         "users",
         metadata,
-        Column("id", BigInteger, primary_key=True),
+        Column("id", Integer, primary_key=True),
         Column("name", String),
     )
 
@@ -31,9 +31,9 @@ class Home(Model):
     __table__ = Table(
         "homes",
         metadata,
-        Column("id", BigInteger, primary_key=True),
+        Column("id", Integer, primary_key=True),
         Column("name", String),
-        Column("owner_id", BigInteger, ForeignKey("users.id")),
+        Column("owner_id", Integer, ForeignKey("users.id")),
     )
 
     __associations__ = Associations(BelongsTo("owner", User, "owner_id"))
@@ -43,9 +43,9 @@ class Pet(Model):
     __table__ = Table(
         "pets",
         metadata,
-        Column("id", BigInteger, primary_key=True),
+        Column("id", Integer, primary_key=True),
         Column("name", String),
-        Column("owner_id", BigInteger, ForeignKey("users.id")),
+        Column("owner_id", Integer, ForeignKey("users.id")),
     )
 
     __associations__ = Associations(BelongsTo("owner", User, "owner_id"))
